refactor: created XMLHelpers directory  - moved ByteToImageConverter here
add: implemented IngredientTemplateDataSelector
