+++ RecipeDetailedView.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:xml-Helpers="clr-namespace:CookNook.Model"
+             xmlns:xml-Helpers="clr-namespace:CookNook.XMLHelpers"
              x:Class="CookNook.RecipeDetailedView"
              Title="{Binding Name}"
              BackgroundColor="#202c39">
